+++ packages/cs-shared/CHANGELOG.md
@@ -15,6 +15,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [5.6.1] - 2021-04-02
+### Added
+ -  new property `UBEntity.overriddenBy` - comma separated model names where entity is overridden
+
+### Changed
+ - **BREAKING** for overridden entities `UBEntity.modelName` property now contains an original model name instead of
+  last override model name
+
 ## [5.6.0] - 2021-03-25
 ### Fixed
  - CustomRepository.fromUbql: prevents overriding of a default `method` in case Repository is created from JSON without method property